Split figures in axisartist and axisgrid1
