+++ doc/mpl_toolkits/axes_grid1/figures/simple_axes_divider2.py
@@ -1,4 +1,4 @@
-import mpl_toolkits.axes_grid.axes_size as Size
+import mpl_toolkits.axes_grid1.axes_size as Size
 from mpl_toolkits.axes_grid import Divider
 import matplotlib.pyplot as plt
