+++ doc/mpl_toolkits/axisartist/figures/axis_direction_demo_step01.py
