+++ doc/mpl_toolkits/axisartist/figures/demo_axis_direction.py
