+++ doc/mpl_toolkits/axisartist/figures/demo_ticklabel_alignment.py
