+++ doc/mpl_toolkits/axisartist/figures/simple_axis_direction01.py
