+++ doc/mpl_toolkits/axisartist/figures/simple_axis_pad.py
