+++ doc/mpl_toolkits/axisartist/figures/simple_axisartist1.py
